refactor(editor): move video url input to own comp
